requirements.txt
fastapi
uvicorn
langchain>=0.2.0
langchain-community
langchain-huggingface
chromadb>=0.5.0
sentence-transformers
pydantic>=2.7.0
python-multipart
pymupdf
langchain-groq
python-dotenv
langchain-mongodb
pymongo